+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The CLI snap version uses the `$SNAP_USER_COMMON` directory for config by default, so that it is preserved between revisions.
 - Defer events subscriptions until there is actual interest for events.
 - End device creation form with wizard in the Console.
+- Gateway connection stats are now stored in a single key.
 
 ### Deprecated